refactor(runtime,threads): 按模块拆分 runtime 与 thread actions 大文件
- 将 runtime 的 process diagnostics、observability 与 pid snapshot 逻辑抽离到独立模块，保留 RuntimeManager facade 与既有测试入口。

- 将 useThreadActions 的 shared/session mutation 动作抽离到 sessionActions 模块，并提取测试共享 harness，保持外部 hook contract 不变。

- 将三处超阈值文件降回 large-file hard gate 以内：runtime/mod.rs 2620 行、useThreadActions.ts 2932 行、useThreadActions.test.tsx 2995 行。

验证：

- npm run check:large-files

- npm run typecheck

- npx vitest run src/features/threads/hooks/useThreadActions.test.tsx

- cargo test --manifest-path src-tauri/Cargo.toml runtime::tests:: -- --nocapture
